feat: Mejorar cálculo del plan de producción y la integridad del reporte
- Modifica el tipo de dato de la propiedad 'Plan' a 'float' en `ProductionReportDto` para permitir valores decimales.
- Refactoriza el `ProductionReportService` para asegurar que el reporte de producción incluya todas las horas del rango solicitado, incluso si no hubo producción en una hora específica.
- Implementa el cálculo de la producción planificada ('Plan') por hora, considerando los tiempos de paro y el tiempo neto de la pieza.
- Optimiza la búsqueda de datos de producción y paros utilizando diccionarios para mejorar el rendimiento.

+++ upmDomain/DomainProductionReport/ProductionReportDto.cs
@@ -19,7 +19,7 @@ namespace upmDomain.ProductionReport
     {
         public DateTime Time { get; set; }
         public int Production { get; set; }
-        public int Plan { get; set; }
+        public float Plan { get; set; }
         public float Efectivity
         {
             get
